second ver

+++ workshop.h
@@ -13,7 +13,7 @@ void printArray(double* array, int size){
 }
 
 double arrayMax(double* array, int size){
-    int max = array[0];
+    double max = array[0];
     for (int i = 0; i<size; i++){
         if (max<array[i]){
             max = array[i];
